fix after review Update RedisConfig Create FollowerEventConfig
