--- src/main/java/faang/school/analytics/mapper/EventMapper.java
@@ -1,8 +1,0 @@
-package faang.school.analytics.mapper;
-
-import faang.school.analytics.dto.AbstractEventDto;
-import faang.school.analytics.model.AnalyticsEvent;
-
-public interface EventMapper {
-    AnalyticsEvent toEntity(AbstractEventDto abstractEventDto);
-}
